pyproject.toml
[tool.poetry]
name = "farm riders python"
version = "0.1.0"
description = ""
authors = ["dakohhh <wiizzydreadmill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.113.0"
bcrypt = "^4.2.0"
PyJWT = "^2.9.0"
slowapi = "^0.1.9"
beanie = "^1.26.0"
fastapi-mail = "^1.4.1"
Jinja2 = "^3.1.4"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
email-validator = "^2.2.0"
pydantic-settings = "^2.4.0"
secure = "^0.3.0"
certifi = "^2024.8.30"
typing-extensions = "^4.12.2"
mypy = "^1.11.2"
mongoengine = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
black = "^24.8.0"
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
skip-string-normalization = true
skip-magic-trailing-comma = true
verbose = true
fast = true

